+++ packages/rx_bloc_cli/lib/src/extensions/arg_results_extensions.dart
@@ -1,7 +1,6 @@
 import 'dart:io';
 
 import 'package:args/args.dart';
-import 'package:rx_bloc_cli/src/extensions/string_extensions.dart';
 import 'package:rx_bloc_cli/src/models/errors/command_usage_exception.dart';
 import 'package:rx_bloc_cli/src/models/realtime_communication_type.dart';
 
@@ -33,8 +32,7 @@ extension ArgumentsValueReader on ArgResults {
   /// Reads a string from the parsed values
   String readString(CommandArguments argument) {
     if (argument.type != ArgumentType.string) {
-      throw UnsupportedError(
-          '${argument.name} is not of type String');
+      throw UnsupportedError('${argument.name} is not of type String');
     }
     return (this[argument.name] as String?) ?? argument.defaultValue();
   }
@@ -42,11 +40,9 @@ extension ArgumentsValueReader on ArgResults {
   /// Reads a boolean from the parsed values
   bool readBool(CommandArguments argument) {
     if (argument.type != ArgumentType.boolean) {
-      throw UnsupportedError(
-          '${argument.name} is not of type bool');
+      throw UnsupportedError('${argument.name} is not of type bool');
     }
-    return (this[argument.name] as String?)?.toBool() ??
-        argument.defaultValue();
+    return this[argument.name] as bool;
   }
 
   /// Reads a realtime communication type from the parsed values
@@ -54,7 +50,7 @@ extension ArgumentsValueReader on ArgResults {
       CommandArguments argument) {
     if (argument.type != ArgumentType.realTimeCommunicationEnum) {
       throw UnsupportedError(
-          '${argument.name} is not of type RealtimeEnumType');
+          '${argument.name} is not of type RealtimeCommunicationType');
     }
     final value = this[argument.name] as String?;